+++ spring-breeze-erp-test/back/src/main/java/com/sb/erp/resv/service/ReservationService.java
@@ -13,7 +13,7 @@ public interface ReservationService {
     // 자원 예약 전체 갯수
     int getResvCount(ResvSearchRequest search);
 
-    ResvResponse getResvDetail(int revId);
+    ResvResponse getResvDetail(long revId);
 
     int insert(ResvRequest ResvDto);
 
@@ -25,7 +25,7 @@ public interface ReservationService {
     StatsResvResponse countByStats(ResvSearchRequest search);
 
     // 예약 관리에서 예약 된 자원이 있는지 확인
-    int countReservationsByResourceId(int resId);
+    int countReservationsByResourceId(long resId);
     
     int updateApprove(ResvRequest resvDto);
     int updateReject(ResvRequest resvDto);
